Finish moving parser, fix up license, catch another place we should use this.
