+++ dspace-api/src/main/java/org/dspace/util/package-info.java
@@ -1,0 +1,3 @@
+/** Generally useful classes, not particularly tied to any other package. */
+
+package org.dspace.util;
